+++ components/internal/forms/Input/react.js
@@ -1,0 +1,32 @@
+import React, { Component } from "react";
+import s from "./style.module.scss";
+import classNames from "classnames/bind";
+import { mX } from "@masonite/svg-icons";
+
+const cx = classNames.bind(s);
+
+export default class Input extends Component {
+  constructor(props) {
+    super(props);
+  }
+
+  render() {
+    const { error, variant } = this.props;
+
+    return (
+      <div className={cx("input-wrapper")}>
+        <input
+          className={cx("input", {
+            "input--dark": variant === "dark",
+            "input--error": error
+          })}
+          type="text"
+        />
+        <div
+          className={cx(["input-icon", { "input-icon--error": error }])}
+          dangerouslySetInnerHTML={{ __html: mX }}
+        />
+      </div>
+    );
+  }
+}