+++ packages/external-ui-react/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Text input
+
+### Changed
+
+- Button small variant text size
+
+### Fixed
+
+- `typography.scss` no longer includes font import as it is variables only.
+
 ## [1.4.0] - 2018-12-31
 
 ### Remove